refactor: move the rest of the things in pages to general module
