+++ src/AppLayout.tsx
@@ -10,7 +10,7 @@ import { useActivityHook } from './modules/discovery/hooks/useActivityHook'
 import { PlatformNavBar } from './modules/navigation/platformNavBar/PlatformNavBar'
 import { InfoBanner } from './modules/notification/InfoBanner'
 import { NoticeBanner } from './modules/notification/NoticeBanner'
-import { LoadingPage } from './pages/loading'
+import { LoadingPage } from './modules/general/loading/index.tsx'
 import { dimensions, ID } from './shared/constants'
 import { useLayoutAnimation } from './shared/hooks'
 import { useCountriesData } from './shared/hooks/graphqlState/useCountriesData.ts'
