+++ src/config/routes/routes.tsx
@@ -6,7 +6,7 @@ import { ActivityDirection } from '@/modules/discovery/pages/activity/components
 import { App } from '../../App'
 import { AppLayout } from '../../AppLayout'
 import { ExternalAuthSuccess, FailedAuth } from '../../modules/auth'
-import { NotAuthorized, NotFoundPage, NotFoundProject } from '../../pages/fallback'
+import { NotAuthorized, NotFoundPage, NotFoundProject } from '../../modules/general/fallback'
 import { __production__, getPath, PathName } from '../../shared/constants'
 import { ErrorBoundary } from './components/ErrorBoundary'
 import { renderPrivateRoute } from './components/PrivateRoute'
@@ -39,7 +39,7 @@ const ProfilePage = () => import('../../modules/profile')
 
 const ProfileSettingsIndex = () => import('../../modules/profile/pages/profileSettings')
 
-const Badges = () => import('../../pages/badges/BadgesPage')
+const Badges = () => import('../../modules/general/badges/BadgesPage')
 
 export const platformRoutes: RouteObject[] = [
   {
