+++ src/modules/general/badges/AccordionPanel.tsx
