+++ src/modules/general/badges/BadgesFAQ.tsx
