--- src/pages/otp/components/index.ts
@@ -1,2 +1,0 @@
-export * from './ReceiveOneTimePassword'
-export * from './VerifyOneTimePassword'
