--- src/pages/otp/index.ts
@@ -1,1 +1,0 @@
-export * from './VerifyYourEmail'
